+++ lib/widgets/value_observable_builder.dart
@@ -9,13 +9,11 @@ class ValueObservableBuilder<T> extends StatefulWidget {
     @required this.builder,
     @required this.stream,
     this.child,
-    this.reuseChild = true,
   }) : super(key: key);
 
   final ValueObservable<T> stream;
   final AsyncChildWidgetBuilder<T> builder;
   final Widget child;
-  final bool reuseChild;
 
   @override
   _ValueObservableBuilderState createState() =>
@@ -23,9 +21,6 @@ class ValueObservableBuilder<T> extends StatefulWidget {
 }
 
 class _ValueObservableBuilderState<T> extends State<ValueObservableBuilder<T>> {
-  AsyncSnapshot<T> _snapshot;
-  Widget _built;
-
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<T>(
@@ -41,12 +36,7 @@ class _ValueObservableBuilderState<T> extends State<ValueObservableBuilder<T>> {
             snapshot.data,
           );
         }
-        // 必要な時(初回・snapshotに差がある時)だけリビルド
-        if (_built == null || !widget.reuseChild || _snapshot != snapshot) {
-          _built = widget.builder(context, snapshot, widget.child);
-        }
-        _snapshot = snapshot;
-        return _built;
+        return widget.builder(context, snapshot, widget.child);
       },
     );
   }